pyproject.toml
[project]
name = "simple-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.2",
    "django-cors-headers>=4.6.0",
    "djangorestframework>=3.15.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "django-environ>=0.11.2",
    "pillow>=11.0.0",
    "django-storages>=1.14.4",
]
